proper resources management in cli test
